refactor ReturnType handling by moving logic to AbstractSorter and introducing ReturnType enum
