+++ Teracura.AbstractSort.Logic/ReturnType.cs
@@ -1,0 +1,9 @@
+﻿namespace Teracura.AbstractSort.Logic;
+
+public enum ReturnType
+{
+    List,
+    Queue,
+    Stack,
+    HashSet,
+}
